@@ -5,16 +5,15 @@
   "main": "./src/index",
   "author": "Justin Spears <speajus@gmail.com>",
   "dependencies": {
-    "@babel/preset-typescript": "^7.1.0",
-    "@types/lodash": "^4.14.119",
-    "@types/node": "^10.12.18",
+    "@babel/preset-typescript": "^7.7.2",
+    "@types/lodash": "^4.14.145",
+    "@types/node": "^12.12.6",
     "@types/promise": "^7.1.30",
-    "@types/puppeteer": "^1.11.1",
-    "@types/react": "^16.7.18",
+    "@types/react": "^16.9.11",
     "source-map-loader": "^0.2.4",
-    "ts-loader": "^5.3.3",
+    "ts-loader": "^6.2.1",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "typescript": "^3.2.2",
+    "typescript": "^3.7.2",
     "webpack-utils": "^1.0.0"
   },
   "keywords": [
