@@ -1,15 +1,22 @@
+export type Project = {
+    name: string,
+    location: string,
+}
+export type LernaOpts = {
+    cwd?: string,
+}
 /**
  * Hard enough for me to think its not a good idea.
  *
  * @param opts
  * @returns {Promise<*>}
  */
-module.exports = (opts) => {
+export default async (opts?: LernaOpts): Promise<Project[]> => {
     try {
         const LsCommand = require('@lerna/list');
         return new Promise((res, onRejected) => {
             const ls = new LsCommand({
-                _   : [],
+                _: [],
                 json: true,
                 ...opts,
                 onRejected,
@@ -28,9 +35,9 @@ module.exports = (opts) => {
         })
     } catch (e) {
         const LsCommand = require('lerna/lib/commands/LsCommand').default;
-        const ls        = new LsCommand(null, opts, opts.cwd || process.cwd());
+        const ls = new LsCommand(null, opts, opts?.cwd ?? process.cwd());
         ls.runPreparations();
-        return Promise.resolve(ls.filteredPackages.map(({_location: location, ...rest}) => ({
+        return Promise.resolve(ls.filteredPackages.map(({_location: location, ...rest}: { _location: string }) => ({
             location,
             ...rest,
         })));
