@@ -1,13 +1,15 @@
-module.exports = function help(optionsManager) {
-    return () => {
-        let str        = '';
-        const aliasMap = {};
+import {OptionsManagerType, OptionType} from "./types";
 
-        optionsManager.forEach((option, key) => {
+export default function help(optionsManager: OptionsManagerType): () => void {
+    return () => {
+        let str = '';
+        const aliasMap: { [key: string]: OptionType[] } = {};
+
+        optionsManager.forEach((option) => {
             if (option.alias) {
-                Object.keys(option.alias).reduce(function (ret, a) {
+                Object.keys(option.alias).forEach(function (a) {
                     (aliasMap[a] || (aliasMap[a] = [])).push(option);
-                }, aliasMap)
+                });
             }
         });
 
@@ -16,7 +18,7 @@ module.exports = function help(optionsManager) {
                 ? 'enabled' : 'disabled'}]\n`;
             if (option.alias) {
                 const keys = Object.keys(option.alias);
-                str        = keys.reduce(function (ret, key) {
+                str = keys.reduce(function (ret: string, key: string) {
                     const val = option.alias[key];
                     return str +=
                         key.length === 1 ? `\v\t-${key}\v\t${val}\n`
